user logout

+++ notes-controller.js
@@ -12,14 +12,26 @@ async function getNotes() {
   return notes;
 }
 
-async function removeNote(id) {
-  await Note.deleteOne({ id });
+async function removeNote(id, owner) {
+  const result = await Note.deleteOne({ id, owner });
+
+  if (result.matchedCount === 0) {
+    throw new Error("No note to delete");
+  }
 
   console.log(chalk.bgGreen.inverse(`Note id:${id} was removed`));
 }
 
-async function modifyNote(noteData) {
-  await Note.updateOne({ id: noteData.id, title: noteData.title });
+async function modifyNote(noteData, owner) {
+  const result = await Note.updateOne(
+    { id: noteData.id, owner },
+    { title: noteData.title }
+  );
+  
+  if (result.matchedCount === 0) {
+    throw new Error("No note to edit");
+  }
+
   console.log(chalk.bgYellow.inverse(`Note id:${noteData.id} was modified`));
 }
